chore: move authorize to UcanAuthorization::redeem, merge test_helpers
- UcanAddress no longer has authorize()
- UcanAuthorization::redeem() exchanges the signed UCAN chain at
  the access service for a presigned URL permit
- test_helpers module merged into helpers
